refactor(cli): move the check group into its own module
Cuts `gpio check` out of cli/main.py into
geoparquet_io/cli/commands/check.py, unchanged, finishing the decomposition
started in #914. main.py goes 1281 -> 189 lines and is now what its own
comment block claims: the root group, the global options and the `skills`
command, plus one `cli.add_command()` per group.

`MultiFileCheckRunner` travels with the group -- it has exactly one user,
`check`, and the six commands that build one. Moving it to `core/` is the
separate "embedded business logic" step, not this one.

The default-subcommand class is constructed in the group module, as `extract`
and `inspect` already do, and renamed `CheckDefaultGroup` to match them. The
factory in cli/_shared.py names every class it builds `_DefaultGroup`, which is
what the CLI surface snapshot pins, so the local name is free.

Eleven imports in main.py were orphaned by the move (`Path`, `_activate_s3`,
`create_default_group`, four decorators, `handle_fix_common`, `CheckProfile`
and the two check impls, `configure_verbose`); each was verified unreferenced
before removal. `GlobAwareCommand` stays for `OptionalIntCommand`, which is
itself dead -- one definition, zero references -- and is left for a follow-up
rather than deleted here.

Seven pre-existing type errors were suppressed by the blanket mypy override on
`geoparquet_io.cli.main` and surface the moment the code moves, so the override
entry follows the code: narrowed to the four codes it actually emits rather
than the errors being fixed here, keeping this a pure move.

No test needed repointing: every `from geoparquet_io.cli.main import check`
still resolves through main.py's own import, and nothing outside the group
referenced `MultiFileCheckRunner` or the command callbacks.

tests/data/cli_surface.json is byte-identical, and all nine `--help` renders
under `gpio check` are unchanged character for character.
